+++ src/service.js
@@ -18,7 +18,7 @@ async function serveModAddStore(logger, port, config) {
   };
 
 
-  const c2c = new CrudToConfig('APPSTORE', undefined, 'source-');
+  const c2c = new CrudToConfig('APP_MANAGER', undefined, 'source-');
   await c2c.login(config.folioInfo);
 
   const app = express();
@@ -38,8 +38,8 @@ async function serveModAddStore(logger, port, config) {
   This is mod-app-manager. Try:
   <ul>
     <li><a href="/admin/health">Health check</a></li>
-    <li><a href="/app-store/config/sources">Configured sources</a></li>
-    <li><a href="/app-store/apps">Apps</a></li>
+    <li><a href="/app-manager/config/sources">Configured sources</a></li>
+    <li><a href="/app-manager/apps">Apps</a></li>
     <li><a href="/htdocs/">Static area</a></li>
     <li><a href="/target/">Generated descriptors</a></li>
   </ul>
@@ -50,28 +50,28 @@ async function serveModAddStore(logger, port, config) {
     res.send('Behold! I live!!');
   });
 
-  app.get('/app-store/apps', async (req, res) => {
+  app.get('/app-manager/apps', async (req, res) => {
     const sourceConfig = await c2c.list();
     console.log('sourceConfig =', sourceConfig);
     returnOrReport(res, () => getAppsFromGitHub(sourceConfig));
   });
 
-  app.get('/app-store/config/sources', async (req, res) => {
+  app.get('/app-manager/config/sources', async (req, res) => {
     returnOrReport(res, () => c2c.list());
   });
 
-  app.post('/app-store/config/sources', async (req, res) => {
+  app.post('/app-manager/config/sources', async (req, res) => {
     const record = req.body;
     returnOrReport(res, () => c2c.add(record), true);
   });
 
-  app.put('/app-store/config/sources/:id', async (req, res) => {
+  app.put('/app-manager/config/sources/:id', async (req, res) => {
     const id = req.params.id;
     const record = req.body;
     returnOrReport(res, () => c2c.update(id, record));
   });
 
-  app.delete('/app-store/config/sources/:id', async (req, res) => {
+  app.delete('/app-manager/config/sources/:id', async (req, res) => {
     const id = req.params.id;
     returnOrReport(res, () => c2c.delete(id));
   });