+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [1.2.0](https://github.com/MikeTaylor/mod-app-manager/tree/v1.2.0) (IN PROGRESS)
 
-* Drive `/app-store/apps` from mod-configuration instead of hardwired config
+* Drive `/app-manager/apps` from mod-configuration instead of hardwired config
 * Write Okapi proxying entries for config CRUD WSAPI
 * Add permissions for config CRUD, including permissions for mod-configuration
 * Add instructions for runing under Okapi.